Default pipeline to run on the HiCDOC analysis.

Runs the default filtering, normalization, and computational steps on a HiCDOCDataSet. To learn more about HiCDOC, browse the vignette: browseVignettes(package = "HiCDOC").

HiCDOC(object, parallel = FALSE)

Arguments

object

A HiCDOCDataSet.

parallel

Whether or not to parallelize each step. Defaults to FALSE.

Value

A HiCDOCDataSet with all slots filled.

Details

HiCDOC pipeline

The HiCDOC pipeline has seven steps: \

Three filtering steps: filterSmallChromosomes to filter out small chromosomes filterWeakPositions to filter out weak positions with very few interactions filterSparseReplicates to filter out sparse replicates with many null interactions Three normalization steps: normalizeTechnicalBiases to normalize technical biases in each replicates normalizeBiologicalBiases to normalize biological biases in each replicate normalizeDistanceEffect to normalize the distance effect in each chromosome One computational step: detectCompartments to detect compartments in each condition and find significant changes between conditions.

Parallel processing

The parallel version of HiCDOC uses the BiocParallel package. Before to call the function in parallel you should specify the parallel parameters such as:

On Linux: multiParam <- BiocParallel::MulticoreParam(workers = 10)

On Windows: multiParam <- BiocParallel::SnowParam(workers = 10)

And then you can register the parameters to be used by BiocParallel: \ BiocParallel::register(multiParam, default = TRUE)

You should be aware that using MulticoreParam, reproducibility of the detectCompartments function using a RNGseed may not work. See this issue for more details.

See also

Examples

data(exampleHiCDOCDataSet)
# Default HiCDOC pipeline
# Not printing loess warnings for example purpose. 
# Results should be inspected if there is any.
suppressWarnings(
object <- HiCDOC(exampleHiCDOCDataSet)
)
#> Keeping chromosomes with at least 100 positions.
#> Kept 3 chromosomes: X, Y, Z
#> Removed 1 chromosome: W
#> Keeping replicates filled with at least 30% non-zero interactions.
#> 
#> Removed interactions matrix of chromosome X, condition 1, replicate R2 filled at 2.347%.
#> Removed 1 replicate in total.
#> Keeping positions with interactions average greater or equal to 1.
#> Chromosome X: 2 positions removed, 118 positions remaining.
#> Chromosome Y: 3 positions removed, 157 positions remaining.
#> Chromosome Z: 0 positions removed, 200 positions remaining.
#> Removed 5 positions in total.
#> Normalizing technical biases.
#> Chromosome X: normalizing biological biases.
#> Chromosome Y: normalizing biological biases.
#> Chromosome Z: normalizing biological biases.
#> Chromosome X: normalizing distance effect.
#> Chromosome Y: normalizing distance effect.
#> Chromosome Z: normalizing distance effect.
#> Clustering genomic positions.
#> Predicting A/B compartments.
#> Detecting significant differences.

# Equivalent to
if(FALSE){
    object <- filterSmallChromosomes(exampleHiCDOCDataSet)
    object <- filterSparseReplicates(object)
    object <- filterWeakPositions(object)
    object <- normalizeTechnicalBiases(object)
    object <- normalizeBiologicalBiases(object)
    object <- normalizeDistanceEffect(object)
    object <- detectCompartments(object)
}