HiCDOCDataSet S4 class.

Data structure for a Hi-C experiment.

Details

An instance of HiCDOCDataSet describes a Hi-C experiment with slots for path(s) to input file(s), interactions, pipeline parameters defaulting to defaultHiCDOCParameters, and computation results. It can be constructed from 4 different types of data:

Tabular files: see HiCDOCDataSetFromTabular

(m)Cool files: see HiCDOCDataSetFromCool

HiC files: see HiCDOCDataSetFromHiC

HiC-Pro matrices and bed files: see HiCDOCDataSetFromHiCPro An example HiCDOCDataSet is also available, see exampleHiCDOCDataSet.

Slots

input

A vector of path(s) to input file(s).

parameters

A list of parameters used for filtering, normalization, and prediction of compartments.

interactions

A tibble of interactions.

chromosomes

A vector of names of chromosomes.

conditions

A vector of names of conditions, repeated along the replicates.

replicates

A vector of names of replicates, repeated along the conditions.

positions

A tibble of positions and their corresponding bin.

binSize

The resolution: computed bin size (span of each bin in number of bases).

totalBins

A list of the number of bins in each chromosome.

weakBins

A list of weak bins that are filtered out in each chromosome.

validConditions

A list of non-sparse valid conditions, repeated along the valid replicates in each chromosome.

validReplicates

A list of non-sparse valid replicates, repeated along the valid conditions in each chromosome.

compartments

A tibble of the A or B compartment of each bin in each condition.

concordances

A tibble of the concordance of each bin in each replicate.

differences

A tibble of detected compartment differences between conditions.

comparisons

A tibble of comparisons ??? TODO

distances

A tibble of the distances to centroids of each bin in each replicate.

centroids

A tibble of centroids in each chromosome and condition.

selfInteractionRatios

A tibble of differences between self interaction and other interactions for each bin in each replicate.

See also