Methods to access a HiCDOCDataSet components.

Retrieve information and results from a HiCDOCDataSet.

# S4 method for class 'HiCDOCDataSet'
chromosomes(object)

# S4 method for class 'HiCDOCDataSet'
sampleConditions(object)

# S4 method for class 'HiCDOCDataSet'
sampleReplicates(object)

# S4 method for class 'HiCDOCDataSet'
compartments(object, passChecks = TRUE)

# S4 method for class 'HiCDOCDataSet'
differences(object, threshold = NULL)

# S4 method for class 'HiCDOCDataSet'
concordances(object, passChecks = TRUE)

# S4 method for class 'HiCDOCDataSet'
show(object)

Arguments

object

a HiCDOCDataSet object

passChecks

logical. Display only the concordances/compartments for the chromosomes passing sanity checks.

threshold

a numeric value between 0 and 1. If no threshold, all the differences will be printed even the non significant ones. Otherwise the differences printed are filtered to show the ones with an adjusted p-value <= threshold.

Value

A character vector (for chromosomes, sampleConditions, sampleReplicates), or a GRanges object (for compartments, concordances, differences).

Functions

chromosomes(): Retrieves the vector of chromosome names.

sampleConditions(): Retrieves the vector of condition names, one for each sample.

sampleReplicates(): Retrieves the vector of replicate names, one for each sample.

compartments(): Retrieves a GenomicRange of the compartment of every position in every condition.

differences(): Retrieves a GenomicRange of the significant compartment differences between conditions, and their p-values.

concordances(): Retrieves a GenomicRange of the concordance (confidence in assigned compartment) of every position in every replicate.

Examples

# Load an example dataset already processed 
# (i.e. after the detection of compartments)
data(exampleHiCDOCDataSetProcessed)

exampleHiCDOCDataSetProcessed
#> Object of class 'HiCDOCDataSet'
#> 
#> - Inputs:
#>   ../example.tsv
#> 
#> - Chromosomes:
#>   X, Y, Z
#> 
#> - Replicates:
#>   condition 2, replicate R2
#>   condition 1, replicate R1
#>   condition 2, replicate R1
#>   condition 1, replicate R2
#>   condition 3, replicate R2
#>   condition 3, replicate R1
#>   condition 1, replicate R3
#> 
#> - Parameters:
#>   smallChromosomeThreshold = 100
#>   sparseReplicateThreshold = 0.3
#>   weakPositionThreshold = 1
#>   cyclicLoessSpan = NA
#>   loessSampleSize = 20000
#>   kMeansDelta = 1e-04
#>   kMeansIterations = 50
#>   kMeansRestarts = 20
#>   PC1CheckThreshold = 0.75
#> 
#> - Methods:
#>   chromosomes(object)
#>   sampleConditions(object)
#>   sampleReplicates(object)
#>   compartments(object)
#>   differences(object)
#>   concordances(object)
#>   parameters(object)
#>   parameters(object) <- list()
#> 
chromosomes(exampleHiCDOCDataSetProcessed)
#> [1] "X" "Y" "Z"
sampleConditions(exampleHiCDOCDataSetProcessed)
#> [1] "2" "1" "2" "1" "3" "3" "1"
sampleReplicates(exampleHiCDOCDataSetProcessed)
#> [1] "R2" "R1" "R1" "R2" "R2" "R1" "R3"
compartments(exampleHiCDOCDataSetProcessed)
#> GRanges object with 1025 ranges and 6 metadata columns:
#>          seqnames      ranges strand |     index condition compartment
#>             <Rle>   <IRanges>  <Rle> | <numeric>  <factor>    <factor>
#>      [1]        X       1-137      * |        81         1           A
#>      [2]        X       1-137      * |        81         2           A
#>      [3]        X       1-137      * |        81         3           A
#>      [4]        X     138-274      * |        82         1           A
#>      [5]        X     138-274      * |        82         2           A
#>      ...      ...         ...    ... .       ...       ...         ...
#>   [1021]        Z 26716-26852      * |       556         1           A
#>   [1022]        Z 26853-26989      * |       557         1           A
#>   [1023]        Z 26990-27126      * |       558         1           A
#>   [1024]        Z 27127-27263      * |       559         1           A
#>   [1025]        Z 27264-27400      * |       560         1           A
#>          centroid.check PC1.check assignment.check
#>               <logical> <logical>        <logical>
#>      [1]           TRUE      TRUE             TRUE
#>      [2]           TRUE      TRUE             TRUE
#>      [3]           TRUE      TRUE             TRUE
#>      [4]           TRUE      TRUE             TRUE
#>      [5]           TRUE      TRUE             TRUE
#>      ...            ...       ...              ...
#>   [1021]           TRUE      TRUE             TRUE
#>   [1022]           TRUE      TRUE             TRUE
#>   [1023]           TRUE      TRUE             TRUE
#>   [1024]           TRUE      TRUE             TRUE
#>   [1025]           TRUE      TRUE             TRUE
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
differences(exampleHiCDOCDataSetProcessed)
#> GRanges object with 54 ranges and 10 metadata columns:
#>        seqnames      ranges strand |     index condition.1 condition.2
#>           <Rle>   <IRanges>  <Rle> | <numeric>    <factor>    <factor>
#>    [1]        X   1645-1781      * |        93           1           2
#>    [2]        X   1645-1781      * |        93           2           3
#>    [3]        X   1782-1918      * |        94           1           2
#>    [4]        X   1782-1918      * |        94           2           3
#>    [5]        X   1919-2055      * |        95           1           2
#>    ...      ...         ...    ... .       ...         ...         ...
#>   [50]        Y 14660-14796      * |       308           2           3
#>   [51]        Y 14797-14933      * |       309           1           3
#>   [52]        Y 14797-14933      * |       309           2           3
#>   [53]        Y 14934-15070      * |       310           1           3
#>   [54]        Y 14934-15070      * |       310           2           3
#>           pvalue pvalue.adjusted direction significance centroid.check
#>        <numeric>       <numeric>  <factor>  <character>      <logical>
#>    [1]         0               0      A->B         ****           TRUE
#>    [2]         0               0      B->A         ****           TRUE
#>    [3]         0               0      A->B         ****           TRUE
#>    [4]         0               0      B->A         ****           TRUE
#>    [5]         0               0      A->B         ****           TRUE
#>    ...       ...             ...       ...          ...            ...
#>   [50]         0               0      B->A         ****           TRUE
#>   [51]         0               0      B->A         ****           TRUE
#>   [52]         0               0      B->A         ****           TRUE
#>   [53]         0               0      B->A         ****           TRUE
#>   [54]         0               0      B->A         ****           TRUE
#>        PC1.check assignment.check
#>        <logical>        <logical>
#>    [1]      TRUE             TRUE
#>    [2]      TRUE             TRUE
#>    [3]      TRUE             TRUE
#>    [4]      TRUE             TRUE
#>    [5]      TRUE             TRUE
#>    ...       ...              ...
#>   [50]      TRUE             TRUE
#>   [51]      TRUE             TRUE
#>   [52]      TRUE             TRUE
#>   [53]      TRUE             TRUE
#>   [54]      TRUE             TRUE
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
concordances(exampleHiCDOCDataSetProcessed)
#> GRanges object with 2089 ranges and 5 metadata columns:
#>          seqnames      ranges strand |     index condition replicate
#>             <Rle>   <IRanges>  <Rle> | <numeric>  <factor>  <factor>
#>      [1]        X       1-137      * |        81         1        R1
#>      [2]        X       1-137      * |        81         1        R3
#>      [3]        X       1-137      * |        81         2        R2
#>      [4]        X       1-137      * |        81         3        R1
#>      [5]        X       1-137      * |        81         3        R2
#>      ...      ...         ...    ... .       ...       ...       ...
#>   [2085]        Z 26990-27126      * |       558         1        R3
#>   [2086]        Z 27127-27263      * |       559         1        R2
#>   [2087]        Z 27127-27263      * |       559         1        R3
#>   [2088]        Z 27264-27400      * |       560         1        R2
#>   [2089]        Z 27264-27400      * |       560         1        R3
#>          compartment concordance
#>             <factor>   <numeric>
#>      [1]           A  -0.0142239
#>      [2]           A  -0.0135911
#>      [3]           B  -0.0201974
#>      [4]           B  -0.0164078
#>      [5]           B  -0.0162860
#>      ...         ...         ...
#>   [2085]           A  -0.0205039
#>   [2086]           A  -0.0196943
#>   [2087]           A  -0.0246548
#>   [2088]           A  -0.0205092
#>   [2089]           A  -0.0208321
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths