Access the parameters of a HiCDOCDataSet.

Retrieves or sets parameters used for filtering, normalization, and prediciton of compartments.

defaultHiCDOCParameters

# S4 method for class 'HiCDOCDataSet'
parameters(object)

# S4 method for class 'HiCDOCDataSet'
parameters(object) <- value

Arguments

object

A HiCDOCDataSet.

value

a named list containing the names and valued of the parameters to change (see Details).

Details

A HiCDOCDataSet's parameters are automatically set to default values retrieved from defaultHiCDOCParameters. They are accessed by filtering, normalization, and compartment detection functions. If those functions are called with custom arguments, the object's parameters are updated to record the actual parameters used. If the object's parameters are customized before calling the functions, the custom parameters will be used.

See filterSmallChromosomes, filterSparseReplicates, filterWeakPositions, normalizeDistanceEffect, and detectCompartments, for details on how these parameters are used.

All parameters are listed here:

smallChromosomeThreshold

The minimum length (number of positions) for a chromosome to be kept when filtering with filterSmallChromosomes. Defaults to defaultHiCDOCParameters$smallChromosomeThreshold = 0.

sparseReplicateThreshold

The minimum percentage of non-zero interactions for a chromosome replicate to be kept when filtering with filterSparseReplicates. If a chromosome replicate's percentage of non-zero interactions is lower than this value, it is removed. Defaults to defaultHiCDOCParameters$sparseReplicateThreshold = 0.3.

weakPositionThreshold

The minimum average interaction for a position to be kept when filtering with filterWeakPositions. If a position's average interaction with the entire chromosome is lower than this value in any of the replicates, it is removed from all replicates and conditions. Defaults to defaultHiCDOCParameters$weakPositionThreshold = 1.

cyclicLoessSpan

The span for cyclic loess normalization used in normalizeTechnicalBiases. This value is passed to multiHiCcompare::cyclic_loess. Defaults to `NA` indicating that span will be automatically calculated using generalized cross validation. For large dataset, it is highly recommended to set this value to reduce computing time and necessary memory.

loessSampleSize

The number of positions used as a sample to estimate the effect of distance on proportion of interactions when normalizing with normalizeDistanceEffect. Defaults to defaultHiCDOCParameters$loessSampleSize = 20000.

kMeansDelta

The convergence stop criterion for the clustering when detecting compartments with detectCompartments. When the centroids' distances between two iterations is lower than this value, the clustering stops. Defaults to defaultHiCDOCParameters$kMeansDelta = 0.0001.

kMeansIterations

The maximum number of iterations during clustering when detecting compartments with detectCompartments. Defaults to defaultHiCDOCParameters$kMeansIterations = 50.

kMeansRestarts

The amount of times the clustering is restarted when detecting compartments with detectCompartments. For each restart, the clustering iterates until convergence or reaching the maximum number of iterations. The clustering that minimizes inner-cluster variance is selected. Defaults to defaultHiCDOCParameters$kMeansRestarts = 20.

PC1CheckThreshold

The minimum percentage of variance that should be explained by the first principal component of centroids to pass sanity check. Defaults to defaultHiCDOCParameters$PC1CheckThreshold = 0.75.

Examples

data(exampleHiCDOCDataSet)

# Retrieve parameters
parameters(exampleHiCDOCDataSet)
#> $smallChromosomeThreshold
#> [1] 100
#> 
#> $sparseReplicateThreshold
#> [1] 0.3
#> 
#> $weakPositionThreshold
#> [1] 1
#> 
#> $cyclicLoessSpan
#> [1] NA
#> 
#> $loessSampleSize
#> [1] 20000
#> 
#> $kMeansDelta
#> [1] 1e-04
#> 
#> $kMeansIterations
#> [1] 50
#> 
#> $kMeansRestarts
#> [1] 20
#> 
#> $PC1CheckThreshold
#> [1] 0.75
#> 

# Set parameters
parameters(exampleHiCDOCDataSet) <- list("smallChromosomeThreshold" = 50)
parameters(exampleHiCDOCDataSet) <- list(
    "weakPositionThreshold" = 10,
    "kMeansRestarts" = 30
)