Normalize technical biases.

Normalizes technical biases such as sequencing depth by using a cyclic loess to recursively normalize each pair of interaction matrices. Depends on multiHiCcompare.

normalizeTechnicalBiases(object, parallel = FALSE, cyclicLoessSpan = NULL)

Arguments

object

A HiCDOCDataSet.

parallel

Logical. Whether or not to parallelize the processing. Defaults to FALSE

cyclicLoessSpan

A numeric value in between 0 and 1. The span for cyclic loess normalization. This value is passed to multiHiCcompare::cyclic_loess. Defaults to NULL, NULL indicates that the value of parameters(object)$cyclicLoessSpan will be used. If this value is NA, the span will be automatically calculated using generalized cross validation. **For large dataset, it is highly recommended to set this value to reduce computing time and necessary memory.**

Value

A HiCDOCDataSet with normalized interactions.

Details

Parallel processing

If parallel = TRUE, the function cyclic_loess is launched in parallel mode, using bplapply function. Before to call the function in parallel you should specify the parallel parameters such as:

On Linux: multiParam <- BiocParallel::MulticoreParam(workers = 10)

On Windows: multiParam <- BiocParallel::SnowParam(workers = 10)

And then you can register the parameters to be used by BiocParallel:

BiocParallel::register(multiParam, default = TRUE)

See also

Examples

data(exampleHiCDOCDataSet)
object <- filterSmallChromosomes(exampleHiCDOCDataSet)
#> Keeping chromosomes with at least 100 positions.
#> Kept 3 chromosomes: X, Y, Z
#> Removed 1 chromosome: W
object <- filterSparseReplicates(object)
#> Keeping replicates filled with at least 30% non-zero interactions.
#> 
#> Removed interactions matrix of chromosome X, condition 1, replicate R2 filled at 2.347%.
#> Removed 1 replicate in total.
object <- filterWeakPositions(object)
#> Keeping positions with interactions average greater or equal to 1.
#> Chromosome X: 2 positions removed, 118 positions remaining.
#> Chromosome Y: 3 positions removed, 157 positions remaining.
#> Chromosome Z: 0 positions removed, 200 positions remaining.
#> Removed 5 positions in total.
# Not printing loess warnings for example purpose. 
# Results should be inspected if there is any.
suppressWarnings(
    object <- normalizeTechnicalBiases(object)
)
#> Normalizing technical biases.