Normalize technical biases.

Normalizes technical biases such as sequencing depth by using a cyclic loess to recursively normalize each pair of interaction matrices. Depends on multiHiCcompare.

normalizeTechnicalBiases(object, parallel = TRUE)

Arguments

object
A HiCDOCDataSet.

parallel
Whether or not to parallelize the processing. Defaults to TRUE.

Value

A HiCDOCDataSet with normalized interactions.

Details

Parallel processing

If parallel=TRUE, the function cyclic_loess is launched in parallel mode, using bplapply function. Before to call the function in parallel you should specify the parallel parameters such as:

On Linux: multiParam <- BiocParallel::MulticoreParam(workers = 10)

On Windows: multiParam <- BiocParallel::SnowParam(workers = 10)

And then you can register the parameters to be used by BiocParallel:

BiocParallel::register(multiParam, default = TRUE)

See also

Examples

data(exampleHiCDOCDataSet)
object <- exampleHiCDOCDataSet
object <- filterSparseReplicates(object)

#> Keeping replicates filled with at least 30% non-zero interactions.

#> Removed interactions matrix of chromosome W, condition 2, replicate R1 filled at 0%.

#> Removed interactions matrix of chromosome X, condition 1, replicate R2 filled at 2.347%.
#> Removed interactions matrix of chromosome X, condition 2, replicate R1 filled at 0%.

#> Removed interactions matrix of chromosome Z, condition 1, replicate R1 filled at 0%.
#> Removed interactions matrix of chromosome Z, condition 2, replicate R1 filled at 0%.
#> Removed interactions matrix of chromosome Z, condition 2, replicate R2 filled at 0%.
#> Removed interactions matrix of chromosome Z, condition 3, replicate R1 filled at 0%.
#> Removed interactions matrix of chromosome Z, condition 3, replicate R2 filled at 0%.

#> Removed 8 replicates in total.

object <- filterWeakPositions(object)

#> Keeping positions with interactions average greater or equal to 1.

#> Chromosome W: 0 positions removed, 80 positions remaining.

#> Chromosome X: 2 positions removed, 118 positions remaining.

#> Chromosome Y: 3 positions removed, 157 positions remaining.

#> Chromosome Z: 0 positions removed, 200 positions remaining.

#> Removed 5 positions in total.

object <- normalizeTechnicalBiases(object)

#> Normalizing technical biases.